requirements.txt
Django==1.7
Pillow==2.5.3

dj-database-url==0.3.0
django-autoslug==1.7.2
django-bootstrap3==4.11.0
django-braces==1.4.0
django-debug-toolbar==1.2.1
django-extensions==1.3.11
django-imagekit==3.2.2
django-pipeline==1.3.25
django-reversion==1.8.2

psycopg2==2.5.4
python-decouple==2.3
Unipath==1.0
waitress==0.8.9

model-mommy==1.2.2
coverage==3.7.1
python-coveralls==2.4.2
